requirements.txt
altgraph==0.17.4
mysql-connector-python==9.4.0
packaging==25.0
pydub==0.25.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pyttsx3==2.99
shiboken6==6.9.1
